tests/futility/test_update: Fix --sys_props argument
The 0x10001 value is supposed to be for the tpm_fwver property, which is
the second element of the --sys_props argument. The error was introduced
in CL:1189249, but has never been noticed because the test case happens
to still pass with the incorrectly specified argument.

BUG=none
TEST=make runfutiltests -j
BRANCH=none

(cherry picked from commit e56f3686526c54c45bf0044ab4068387a92e1cef)

Change-Id: Ia05d441daa8848531892f686575ecd813e7ff916
Original-Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5822915
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
GitOrigin-RevId: e56f3686526c54c45bf0044ab4068387a92e1cef
Cr-Build-Id: 8737902702090947393
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8737902702090947393
Copybot-Job-Name: android-main-vboot_reference-copybot-downstream