signing: miniOS signing in docker.
For some reason, lsblk reports null partition types in docker, even in
privileged
(https://forums.docker.com/t/why-does-lsblk-report-partition-as-null-from-within-a-docker-container/139393).
Since we are already hard-coding the miniOS loop devices, stop checking
for the partition type explicitly.

This commit also removed the check for the private key on disk because
the key lives in Cloud KMS, and adds a newline to an info statement that
was getting concatenated to the previous log message.

BUG=b:363466852
BRANCH=None
TEST=manual

Change-Id: I044c626dfac5ae7115b9f6b7e65e8303595544fb
Signed-off-by: Benjamin Shai <bshai@google.com>