futility: updater: Process custom label as standard models
One advantage of switching to signer_config.csv and get rid of
setvars.sh is we finally can know what custom label devices are
supported.

In `setvars.sh` or the pre-Unified-Build single package format, we have
to scan the 'keyset/' folder to figure out what custom label tags may be
available. But when using signer_config.csv, all the custom label names
are already included in the signer information.

As a result, we can finally remove the hacky way of patching custom
label devices. One downside is we can no longer support custom label in
non-Unified-Build systems, but all active Chromebooks are now Unified
Build so that is not a problem.

This is also the first step to remove the custom label logic from the
firmware updater - preparation to let crosid handle custom label
matching for us. This also implies `--signature_id` is now equivalent to
`--model`.

BUG=b:251040363
TEST=make runfutiltests -j
BRANCH=none

Change-Id: Ib4c62ba4d01b276472575625ea19397fc21b1589
Signed-off-by: Hung-Te Lin <hungte@chromium.org>