futility: updater: Remove 'allow_empty_custom_label_tag' quirk
The 'allow_empty_custom_label_tag' quirk was introduced for non-unified
builds to support custom label devices without a tag (customization_id)
set. The only device use it is Hana, which is now an unified build.

After CL:5872452, we are now able to directly load the default keys
using the information provided by `signer_config.csv` even for Hana (and
removed where the quirk was previously used), so this quirk is no longer
needed.

BUG=b:251040363
TEST=make runfutiltests -j
BRANCH=None

Change-Id: I14edd5f89d3518a4ba936a6c4df5e91368d03783
Signed-off-by: Hung-Te Lin <hungte@chromium.org>