futility: updater: Add a new config 'output_only'
There was no easy way to identify if the updater is running in
output-only mode (`--mode=output`). However, some setup process
(especially the custom label tag processing) will want to know if it is
appropriate to read the system firmware. For example, the ODMs and
factories will want to run `--mode=output` in the CrOS SDK to get
the preflash image, and definitely not reading the host system firmware.

After `--signature_id` is deprecated, specifying `--model` without
tags will trigger reading system firmware. As a result, we need to
identify if the updater is running in the output mode, and skip reading
system firmware in that mode. Having the output_only will make it
simpler instead of having the `strcmp(arg->mode, "output")` everywhere.

BUG=b:251040363
TEST=make runfutiltests -j
BRANCH=None

Change-Id: I5181c9a31df7db12e04fdd362fe02ca08be29916
Signed-off-by: Hung-Te Lin <hungte@chromium.org>