host/lib/crossystem: Make CROSSYSTEM_LOCK_PATH configurable
Android (and other systems) do not have `/run/lock` directory, which is
used by crossystem by default to create a lockfile for its `set`
operations. This issue can be circumvented by providing a way to
configure this path via a define and making it configurable by `make`
variable.

For Android the `/data/local/tmp` was picked as it seems to be a default
location for tools for this purpose. (See: Android gdbclient.py
debug_socket and server_remote_path). Although this directory is
persistent across reboots, the lock file is not required to be and path
an be changed later to other one with e.g. tmpfs.

Linux uses /run/lock starting from Filesystem Hierarchy Standard 3.0.
Darwin/Mac use /tmp ad there is no direct equivalent of /run/lock.
Windows uses C:\windows\temp.

IMPORTANT!!!: Windows and Darwin were not tested.

BUG=b:369294243, b:369290629
BRANCH=None
TEST=crossystem dev_boot_usb=1 # On Android

Change-Id: I847f782fb5994185fbaf8f5dd23b60fc74a2c882
Signed-off-by: Jakub Czapiga <czapiga@google.com>