2load_android_kernel: Add fastboot cmdline
Add definition of fastboot_cmdline which is stored on the vendor space
of the misc partition. Parameters from fastboot_cmdline are added to
vboot cmdline only if the FW is in developer mode. To accommodate
potential arguments in the fastboot_cmdline, the space for vboot cmdline
in kernel buffer is extended.

BUG=b:372863538
TEST=Add arguments using fastboot, boot Android and check if arguments
     are present in bootconfig.
BRANCH=firmware-android-15949.B

Change-Id: I786a158a597d0551c89f69c26b9bf80a73459d14
Signed-off-by: Tomasz Michalec <tmichalec@google.com>