futility: updater: Increase try count from 11 to 13 Android Bringup: See http://go/android-fw-sync
Theoretically 11 reboots are enough for firmware update in any
situation. However, a user may power off the device within 1 miniute
after bootup, causing chromeos-setgoodfirmware to not run. The result is
that, fw_result won't be set to "success" in that boot, so fw_try_count
will be further decreased by 1 in the next boot.

Given that the firmware cannot tell if the poweroff is intentional by
the user or unexpected by the system, the best we can do is to take this
edge case into account in the initial fw_try_count value. Increase it
from 11 to 13, so that users will be less unlikely to hit this edge
case. More precisely, they will need to power off the device within 1
minute for 3 consecutive boots after firmware update, in order to hit
this issue.

BUG=b:374893421
TEST=cq
BRANCH=none

(cherry picked from commit 3246e484ca08f2ab29935cb964fa52fc6d24f97b)

Change-Id: Iec21a26b62bdcf6aa6ad44402cd39cebd70c24b8
Original-Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/5970064
Original-Reviewed-by: Julius Werner <jwerner@chromium.org>
Original-Reviewed-by: Hung-Te Lin <hungte@chromium.org>
GitOrigin-RevId: 3246e484ca08f2ab29935cb964fa52fc6d24f97b
Cr-Build-Id: 8732659219451911521
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8732659219451911521
Copybot-Job-Name: android-main-vboot_reference-copybot-downstream