tlcl: Increase TPM buffer size
This change increase the TPM message buffer size from 512 to 1024 in
order to be able to fit new nvmem entry in single read command. Prior
this change the TlclRead(0x3fff0a) would fail, because the space size
was too big to fit into response buffer. With this change the nvmem
content can be read safely.

BUG=b:359340876
BUG=b:354045389
TEST=Verify that nvmem contents are read correctly.
BRANCH=firmware-android-15949.B

Change-Id: Id227acc5a40d47ec81c2e2ddc83503763e0c88f3
Signed-off-by: Bartłomiej Grzesik <bgrzesik@google.com>