futility: updater: Increase try count from 13 to 17
The current try count is 13 for two reasons:

1. As explained in CL:5690663, Rex needs 11 reboots in the worst case.
2. In CL:5970064, we added an extra 2 boots to mitigate the problem of
   abrupt user poweroff during firmware update process.

However, Nissa needs 13 reboots in the worst case.

- Boot 0: Initial reboot with CSE in RW. To allow CSE update, reboot to
          switch to CSE RO.
- Boot 1: Update CSE RW. Reboot to switch to CSE RW.
- Boot 2: GSC needs update. Reset GSC to switch to the other partition.
- Boot 3: CSE is back in RO due to GSC reset. Reboot to switch to RW.
- Boot 4: EC needs update. Reboot to EC RO to allow EC sync.
- Boot 5: CSE is back in RO due to EC reset. Reboot to switch to RW.
- Boot 6: Attempt to perform EC sync, but need display to show the EC
          sync screen. Reboot to init display.
- Boot 7: Perform EC sync. Under EFS2 NO_BOOT mode, reboot to EC RO to
          allow jumping to RW.
- Boot 8: CSE is back in RO due to EC reset. Reboot to switch to RW.
- Boot 9: Perform auxfw sync and reboot to EC RO.
- Boot 10: CSE is back in RO due to EC reset. Reboot to switch to RW.
- Boot 11: Everything is updated, but need to reboot again to unset
           display.
- Boot 12: Boot to kernel, finally!

The only difference between Nissa and Rex is the display init. On Rex,
the display is always initialized for the splash screen. On Nissa, 2
extra reboots are needed to enable and disable display, respectively.

In addition, the 2 boots reserved for poweroff mitigation are actually
not enough, because after the abrupt user poweroff, the CSE will go back
to RO. For example, if the poweroff happens in Boot 9 (before auxfw
sync), then the device will go back to Boot 8, resulting in 2 extra
reboots.

Therefore, to allow 2 cases of abrupt poweroff, we will need extra 4
boots in the try count. Together with Nissa's worst-case scenario, we
will need an initial try count as large as 17.

BUG=b:385224631
TEST=none
BRANCH=none

Change-Id: I4503ef2f519f5368a4697861d1dffac5fbe1cfb5
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>