vboot: modify GptNextKernelEntry function
The GptNextKernelEntry function returns the start & size of the boot
partition. These data are required for booting ChromeOS, but to boot
Android we need a partition suffix. This change modifies
GptNextKernelEntry, so it returns a pointer to GptEntry now. It will
allow to determine which OS is going to be booted by entry type GUID
and simplify fetching data required for a given OS.

BUG=b:384095358
TEST='make runtest' and verify build ChromeOS on redrix
BRANCH=main

Change-Id: I6c3e4555db9f429cd7bcd96e29c8ab661330de55
Signed-off-by: Grzegorz Bernacki <bernacki@google.com>