2api: Separate kernel cmdline and bootconfig
Add pointer to bootconfig in vb2_kernel_params, so
vb2_load_android_kernel may control which parameters are appended to
bootconfig and which to cmdline. It is required for early init
parameters that needs to be present in kernel cmdline before kernel
parses bootconfig.

Reserve space for fastboot bootconfig in misc partition.

BRANCH=firmware-android-15949.B
BUG=b:374092472
TEST=Boot Android and check cmdline and bootconfig content

Cq-Depend: chromium:6179451
Change-Id: I71882aa6e88d8e18a7da60ddb4f63078d294b091
Signed-off-by: Tomasz Michalec <tmichalec@google.com>