cgptlib: Move GptNextKernelEntry declaration
Previously it resided in cgptlib.h. Move it to gpt_misc.h, so that it
can be used in depthcharge. Specifically it'll be used by fastboot
logic to figure out the which slot is active. ("a"/"b")
Since cgptlib.h already includes gpt_misc.h, this change won't affect
existing code.

BRANCH=none
BUG=b:397322409
TEST=emerge-brya vboot_reference libpayload depthcharge

Signed-off-by: Kornel Dulęba <korneld@google.com>
Change-Id: I2f2af87cfc5e7e92532c6a13515f638bbae5d451