futility: updater: Cap try count to 15
We increase the try count to 17 in CL:6164433, in order to allow 2 cases
of abrupt poweroff by the user. However, VB2_NV_TRY_COUNT has only 4
bits in vboot nvdata, so the maximum effective value is 15. The maximum
value is enough for the all existing platforms, including 1 case of
abrupt poweroff.

BUG=none
TEST=cq
BRANCH=none

Change-Id: I83a45a32cae16b46ca5f4d95f7ea5ebcd17cf4f7
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>