sign_gsc_firmware: extend to support signing NT images
As presented NT images are signed with Cloud KMS hosted ECDSA key
using openssl with the PKCS#11 plugin. This covers the case of
semi-prod signing for nightly tests.

C4A HSM signing with both SPX and ECDSA key is a future enhancement,
but provisions are being made for adding both SPX and ECDSA signature
to the image prepared by the builder.

BRANCH=none
BUG=b:395968753, b:400968066
TEST=validated proper ECDSA signing and OT image creation.

Change-Id: I239fd87d09391cc55581cb40dce019168f022b60
Signed-off-by: Vadim Bendebury <vbendeb@google.com>