cgptlib: Move GptPartition and GptPartitionNames declarations
Previously it resided in cgptlib.h. Move it to gpt_misc.h, so that it
can be used in depthcharge. Specifically it'll be used by fastboot logic
to get the name of misc partition. Since cgptlib.h already includes
gpt_misc.h, this change won't affect existing code.

BRANCH=main
BUG=None
TEST=emerge-brya vboot_reference libpayload depthcharge

Change-Id: I22e82fb72cd200977b30dc36f667b31d04e3a155
Signed-off-by: Tomasz Michalec <tmichalec@google.com>