firmware/2lib: Separate EC sync phase 1 from ec_sync API
Currently, the entire EC sync is handled by vb2api_ec_sync API. This
function checks if the EC firmware update is necessary and if so
performs the update and jumps to updated firmware, else just switches
to the already existing RW firmware.

For some ECs, the firmware update operation takes around 15-20s to
complete. This creates a necessity to display SOL for better user
experience. Depthcharge already handles this scenario and enables SOL,
but when EC sync is done early from coreboot, eSOL is not enabled.
In order to display eSOL, there is a need to know whether there is
an actual EC firmware update initiated.

This change separates the EC sync phase 1 from `vb2api_ec_sync()` and
adds a new API `vb2api_ec_sync_phase1()` to perform phase 1 operations.
Another API `vb2api_is_ec_sync_required()` is introduced which can be
used after EC sync phase 1 to check if the EC firmware update is
initiated. This API can be used to decide whether eSOL should be
displayed or not.

BUG=b:412210635
TEST=make runtests; Make sure EC sync operation is not affected in
both coreboot and depthcharge.

Change-Id: Ia9c8581f5f2ffd45570dfc9dea1a7dba193a02b2
Signed-off-by: Pranava Y N <pranavayn@google.com>