futility: updater: Allow --quirks=preserve_me=2 to force preserving ME
Some developers want to flash a firmware from a totally different branch
and the CSME may become totally different (and the CSME on the device
was already locked).

The `--preserve_me` quirk was designed for autoupdate, and in
recovery/factory modes we believe the device should be manually fixed so
the CSME & RO can be compatible. However, having a simple quirk to allow
developers who know their CSME/RO versions are not a problem would be
very helpful.

As a result, this patch extends the existing `--quirks=preserve_me`
so when a >=2 value is assigned then the developers can easily flash
their devices with a new firmware.

BUG=None
TEST=make runfutiltests -j
BRANCH=None

Change-Id: I894a0f9d18e12e3fea40aeda7bae31991bd87789
Signed-off-by: Hung-Te Lin <hungte@chromium.org>