swap_ec_rw: Support legacy EC image
The script only supported the Zephyr EC image, which contains the RW_FW
FMAP section, which describes the effective image. However, the legacy
EC image doesn't have the RW_FW section; it only has the EC_RW section,
which describes the allocated slot.

To determine the effective image size, we need to search for and parse
the image_data header.

This change adds support for the legacy EC image. It searches the
cookies within the image_data header, retrieve the image_data.size,
and then truncate the image accordingly.

BUG=b:418274770
TEST=make runtestscripts
BRANCH=none

Change-Id: If9ed8e143fa211f49d36800d47df0ac03e3e9c7f
Signed-off-by: Wai-Hong Tam <waihong@google.com>