futility/updater: Add FMAP guessing and optimize legacy update
This change adds FMAP location guessing when reading a firmware image based on the provided helper image.
It also introduced the ability to specify what regions need to be read
when loading system firmware, which helped optimize legacy update.

======

The heuristic used here utilizes the fact that usually the system FMAP section can be found at the same offset as the FMAP in the provided helper image, thereby minimizing the amount of data that has to be read by only checking whether the respective memory segment does indeed contain a valid FMAP section.

If guessing fails, the FMAP section will be found by searching the whole flash, as there is no other choice at the moment (potentially a different FMAP guessing attempt could be made using system files that store FMAP location). However, searching flash for FMAP usually is faster than reading the entire image. Therefore, if there are regions that have to be read, FMAP is found without reading the entire image. In the worst case scenario, this will result in the whole image being read, which is inefficient and cannot be optimized.

No changes have been introduced to the process of reading an entire image (when no regions are specified).

======

The legacy firmware update has been optimized by only reading RO_FRID, which is the only section needed during the update. The ability to request only certain regions to be read could be potentially used to optimize try_rw and rw updates, as those do not require the entire image either.

BRANCH=main
BUG=b:431709118
TEST=Ran runfutiltests.
TEST=Tested failure to guess FMAP by artificially changing fmap location of the provided image.

Change-Id: I412ba89f2cea2155804e2687beb1c9675aba3032
Signed-off-by: Ihor Lukianov <ihorlukianov@google.com>