futility/updater: Add FMAP guessing based on helper image
This change adds FMAP location guessing when reading a firmware image based on the provived helper image, as well as the ability to specify what regions need to be read when loading system firmware.

The heuristic used here utilizes the fact that usually the system FMAP section can be found at the same offset as the FMAP in the provived helper image, thereby minimizing the amount of data that has to be read by only checking whether the respective memory segment does indeed contain a valid FMAP section.

If guessing fails, the FMAP section will be found by searching the whole flash, as there is no other choice at the moment (potentially, a different FMAP guess could be made using system files that store FMAP location). However, searching flash for FMAP usually is faster than reading the entire image. Therefore, if there are regions that have to be read, FMAP is found without reading the entire image. In the worst case scenario this might result in the whole image being read.

No changes have been introduced to the process of reading an entire image (when no regions are specified).

BRANCH=main
BUG=None
TEST=None

Change-Id: I412ba89f2cea2155804e2687beb1c9675aba3032
Signed-off-by: Ihor Lukianov <ihorlukianov@google.com>