futility/update: Locate FMAP using image and optimize legacy update
When updating firmware, usually the image provided will have its FMAP
located at the same offset where the system FMAP is in flash. This is
utilized to optimize reading routine by first reading flash where the
FMAP is likely to be found, and only in case of failure is a full flash
search executed.

The legacy firmware update has been optimized by only reading RO_FRID,
which is the only section needed during the update. The ability to
request only certain regions to be read could be potentially used to
optimize try_rw and rw updates, as those do not require the entire
image either.

BRANCH=main
BUG=b:431709118
TEST=Ran runfutiltests.
TEST=Tested failure to guess FMAP by artificially changing fmap
location of the provided image.

Change-Id: I412ba89f2cea2155804e2687beb1c9675aba3032
Signed-off-by: Ihor Lukianov <ihorlukianov@google.com>