futility/updater: Introduce load_system_frid()
Add a new function load_system_frid() to load the system FRID, with the
version numbers trimmed from the returned string. The returned FRID is
the same as the "frid" attribute in the CrOS config "identity" node.

Refactor manifest_detect_model_from_frid() to use the new function.

BUG=b:431694612
TEST=cq
BRANCH=none

Change-Id: I682d95c07f138f035fb1ea419fb99efcf5111855
Signed-off-by: Yu-Ping Wu <yupingso@chromium.org>