crossystem: Remove wpsw_cur fallback to VDAT_INT_HW_WPSW_BOOT
VDAT_INT_HW_WPSW_BOOT has been deprecated a long time ago and is no longer
valid. Using it as a fallback value, which most often is just zero, is
misleading and causes problems.

This code should never be reached on devices with correct configuration,
and if it is, then the user should be informed about the problem instead
of defaulting to the incorrect value.

BRANCH=main
BUG=b:254337014
TEST=CQ

Change-Id: I2e3c1922ba9b767719be6fd7f0ca6f0ba5e753c9
Signed-off-by: Jakub Czapiga <czapiga@google.com>