host: crossystem: gpio: Use correct GPIO controller offset
Some x86 devices have more than one /dev/gpiochipX. Because of that
crossystem was reading the first available pin value instead of the
correct one. To solve this issue this CL makes use of the GPIO
controller name stored in GPIO.3 API pin attribute, extracts the
controller offset and uses it to select the matching /dev/gpiochipX
device. If the controller name does not contain the offset value, then
the code defaults to the first device i.e. /dev/gpiochip0.
The cases when there are more than one controller, but the attribute
does not contain it, are not covered by this implementation.
The ARM code behavior remains unchanged, and still looks for the named
pin.

BRANCH=main
BUG=b:434590349
TEST=`crossystem wpsw_cur` output reflects actual HW WP state
on x86 and ARM64 devices.

Change-Id: I987e75389481ed6ca01003453eec78c1cfce79dc
Signed-off-by: Jakub Czapiga <czapiga@google.com>