common: make optional arg actually optional Android Bringup: See http://go/android-fw-sync
This argument states that it is optional in the documentation, but if
you attempt to run the script without it using `set -u` it complains
that it is unset. Thus, default to empty if it's not provided to avoid
that error.

BRANCH=None
BUG=b:329892770
TEST=manual

(cherry picked from commit de036f432202ab9576e0e0bf101972d8c24573cc)

Change-Id: I179a135d75f6af07db0c3fa6157b9bd09073c45b
Original-Signed-off-by: Benjamin Shai <bshai@google.com>
Original-Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/6806962
Original-Reviewed-by: Madeleine Hardt <hardtmad@google.com>
Original-Commit-Queue: Madeleine Hardt <hardtmad@google.com>
GitOrigin-RevId: de036f432202ab9576e0e0bf101972d8c24573cc
Cr-Build-Id: 8707775259135727393
Cr-Build-Url: https://cr-buildbucket.appspot.com/build/8707775259135727393
Copybot-Job-Name: android-main-vboot_reference-copybot-downstream