futility: Add unit tests for updater utils
Add unit tests for updater utils found in `updater_utils.c`
The tests feature two main files. `test_updater_utils.c` tests code in
`updater_utils.c` file except for the functions working with SERVO, which
are tested in `test_updater_utils_servo.c`. This is necessary because
SERVO will not be usually available during the execution of the tests,
so it has to be emulated by replacing `host_shell` function. However,
that function is also used in the updater utils, so it has to be
replaced in a different runtime.

All public functions in `updater_utils.c` have been added the
`test_mockable` tag (which is basically `__weak`). It is necessary since
without it the tests do not compile due to multiple definitions of
public functions from `updater_utils.c`. It is also the reason why
`is_ap_write_protection_enabled` and `is_ap_write_protection_enabled`
have been stripped of the `inline` tag.

`image-steelix.bin` is the main firmware image file which
is used in these tests.

`images.zip` contains `tests/futility/data_copy/image-steelix.bin`.

`image-missing-*.patch` contain patches which are used to make images
with missing FMAP, RO_FRID, and RW_FWID.

`unit_test_common.h` is a header file with definitions and
other tools which are used in these tests and will be used in the future
ones.

`prepare_data_for_unit_tests.sh` is a script which makes new images,
files, and other data needed by the unit tests.

BRANCH=main
BUG=None
TEST=compile
TEST=verify manually

Change-Id: I29f003f91afeda99a9303e8123b3b49324ab874d
Signed-off-by: Ihor Lukianov <ihorlukianov@google.com>