futility: Add unit tests for updater utils
Add tests for updater utils found in updater_utils.c

image-newer.bin is the main (relatively new) firmware image file which
is used in these tests.

unit_test_common.h is a header file with a bunch of definitions and
other tools which are used in these tests and will be used in the future
ones.

prepare_data_for_unit_tests.sh is a script which makes new images,
files, and other data needed by the unit tests. Creating those during
runtime is the better solution here, as storing that data would take up a few
hundred megabytes of disk space.

The tests feature two main files. test_updater_utils.c tests code in
updater_utils.c file except for the functions working with SERVO, which
are tested in test_updater_utils_servo.c. This is necessary because
SERVO will not be usually available during the execution of the tests,
so it has to be emulated by replacing host_shell() function. However,
that function is also used in the updater utils, so it has to be
replaced in a different runtime.

BRANCH=main
BUG=None
TEST=None

Change-Id: I29f003f91afeda99a9303e8123b3b49324ab874d
Signed-off-by: Ihor Lukianov <ihorlukianov@google.com>