2load_android_kernel: Quick fix for 4KiB misc partition clear
When storage has 4KiB block size, writes to misc partition are not block
aligned. Because of that it is imposible to clear bootonce-bootloader
command when block size is 4KiB.

Use 4KiB buffer when accessing misc partition. This is quick fix for
firmware-android branch. This will be properly fixed on main branch.

BRANCH=firmware-android-15949.B
BUG=b:425282891
TEST=Enter fastboot using "reboot bootloader" on DUT with 4KiB block
     size. Exit fastboot using "reboot" command and boot back to OS.

Change-Id: I9f8f8fbbdc836b8aa4477f5a3cd61028b7b1d2b3
Signed-off-by: Tomasz Michalec <tmichalec@google.com>