Reorder recovery kernel hash updates after MiniOS resigning
The calls to `update_recovery_kernel_hash` for recovery images are moved
to occur after the `resign_minios_kernels` function is executed.

This is to prep it for including various digests/values into the kernel
config at the very end. Allowing for usage of finalized digests to end
up on the kernel config based on any partition.

Bug: chromium:426664945
Test: CQ
Change-Id: Ia230c768cd155812b06c803c02b939b86666b543