Makefile: Introduce stub layer for flashrom access
Currently, we uuse USE_FLASHROM=0/1 flag to decide if we want to build
with flashrom or not. This makes it to be difficult to add alternative
implementations.

This CL introduces a new `FLASHROM_BACKEND` variable in the Makefile to
abstract the flashrom implementation. The supported backends are:
* `lib`: The existing implementation using libflashrom. (i.e.
  flashrom_drv)
* `none`: A stub implementation for builds without flashrom dependency.

Key changes:
* Rename `flashrom_drv.c` to `flashrom_lib.c` for clarity.
* Add `flashrom_stub.c` to provide no-op functions.
* Reworked the Makefile logic to use `FLASHROM_BACKEND`

To maintain backward compatibility, a warning is issued if the
`USE_FLASHROM` flag is set, and its value is translated to the new
backend selection.

The following commit will introduce a `bin` backend, which will execute
the `flashrom` binary as subprocess. Also there will be another final
cleanup commit that remove the `USE_FLASHROM` and all remaining `#ifdef
USE_FLASHROM` directives from the C code.

BUG=b:441639102, b:439499932
BRANCH=none
TEST=Prepare ChromiumOS libflashrom in host.
TEST=make clean && make
TEST=make clean && make FLASHROM_BACKEND=lib
TEST=make clean && make USE_FLASHROM=0

Change-Id: Ibc9826349d2b240758cdd504fc859449ce247484
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>