avb/vboot_avb_ops: Continue without buffer for optional partitions
reserve_buffer tries to allocate/reserve buffer for all partitions even
when they do not exist. This causes failure to load kernel even when the
concerned non-existent partition is not requested to be loaded. Fix this
by updating reserve_buffer to continue when a partition does not exist.

BUG=b:394980221, b:443764205
BRANCH=None
TEST=USE=avb emerge-rauru coreboot libpayload vboot_reference
depthcharge chromeos-bootimage.

Change-Id: Idc6d780f65008abb17620a5a4b04ea1fd4107947
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@google.com>