crossystem: Use libflashrom instead of flashrom CLI for USE_FLASHROM=1
When USE_FLASHROM=1, crossystem currently invokes the flashrom binary
for flash operations via flashrom_read() and flashrom_write(). This is
inconsistent with futility, which already use the libflashrom.

This CL replaces these functions with their libflashrom implementations
flashrom_read_region() and flashrom_write_region(). This unifies the
flash access logic with futility, making it easier to maintain.

Key changes:
* The do_verify parameter in write functions is changed from int to bool
  for correctness.
* The new ...region() functions now require a non-NULL region name. The
  old implementation treated a NULL region as the entire chip; this is
  no longer supported by these functions.

The binary invocation logic in flashrom.c is kept for now and will be
refactored in subsequent CLs.

BUG=b:441639102
BUG=b:445126698
BRANCH=none
TEST=sudo emerge -1 vboot_reference
TEST=sudo USE="-flashrom" emerge -1 vboot_reference
TEST=emerge-corsola vboot_reference, deploy on DUT, check ldd to ensure
     running crossystem with libflashrom
TEST=(on corsola) crossystem fw_try_count
TEST=(on corsola) crossystem fw_try_count=1
TEST=FEATURES=test emerge-corsola vboot_reference
TEST=make clean && USE_FLASHROM=1 make runtests
TEST=make clean && USE_FLASHROM=0 make runtests

Change-Id: I65bf6d620687ca6a1214ed474e451596cb7eb325
Signed-off-by: Hsuan Ting Chen <roccochen@chromium.org>